requirements.txt
mcp[cli]>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
openai>=1.0.0 